implemented alternative min cut (w/ slightly better performance) + moved stuff around + other minor changes
* blend option not yet added to nodes, might cleanup the code a bit first.
* patch_search.py needs to change name, or maybe be separated into different scripts; TBD.
