+++ make_seamless.py
@@ -1,10 +1,7 @@
-from .patch_search import get_generic_find_patch_method, blur_patch_mask
-from .misc.bse_type_aliases import num_pixels
+from .patch_search import get_generic_find_patch_method, get_4way_min_cut_patch
 from .types import UiCoordData
 from math import ceil
 import numpy as np
-
-inf = float('inf')
 
 
 def get_numb_of_blocks_to_fill_stripe(block_size, overlap, dim_length):
@@ -82,108 +79,6 @@ def make_seamless_horizontally(image, block_size, overlap, tolerance, rng: np.ra
         return None
 
     return texture_map
-
-
-def get_min_cut_patch_mask_horizontal(block1, block2, block_size: num_pixels, overlap: num_pixels):
-    """
-    @param block1: block to the left, with the overlap on its right edge
-    @param block2: block to the right, with the overlap on its left edge
-    @return: ONLY the mask (not the patched overlap section)
-    """
-    err = ((block1[:, -overlap:] - block2[:, :overlap]) ** 2).mean(2)
-    # maintain minIndex for 2nd row onwards and
-    min_index = []
-    E = [list(err[0])]
-    for i in range(1, err.shape[0]):
-        # Get min values and args, -1 = left, 0 = middle, 1 = right
-        e = [inf] + E[-1] + [inf]
-        e = np.array([e[:-2], e[1:-1], e[2:]])
-        # Get minIndex
-        min_arr = e.min(0)
-        min_arg = e.argmin(0) - 1
-        min_index.append(min_arg)
-        # Set Eij = e_ij + min_
-        Eij = err[i] + min_arr
-        E.append(list(Eij))
-
-    # Check the last element and backtrack to find path
-    path = []
-    min_arg = np.argmin(E[-1])
-    path.append(min_arg)
-
-    # Backtrack to min path
-    for idx in min_index[::-1]:
-        min_arg = min_arg + idx[min_arg]
-        path.append(min_arg)
-    # Reverse to find full path
-    path = path[::-1]
-    mask = np.zeros((block_size, block_size, block1.shape[2]), dtype=block1.dtype)
-    for i in range(len(path)):
-        mask[i, :path[i] + 1] = 1
-    return mask
-
-
-def get_4way_min_cut_patch(ref_block_left, ref_block_right, ref_block_top, ref_block_bottom,
-                           patch_block, block_size, overlap):
-    # (optional step) blur masks for a more seamless integration ( sometimes makes transition more noticeable, depends )
-    masks_list = []
-
-    has_left = ref_block_left is None
-    has_right = ref_block_right is not None
-    has_top = ref_block_top is not None
-    has_bottom = ref_block_bottom is not None
-
-    if has_left:
-        mask_left = get_min_cut_patch_mask_horizontal(ref_block_left, patch_block, block_size, overlap)
-        mask_left = blur_patch_mask(mask_left, block_size, overlap, has_left, has_right, has_top, has_bottom)
-        masks_list.append(mask_left)
-
-    if has_right:
-        mask_right = get_min_cut_patch_mask_horizontal(np.fliplr(ref_block_right), np.fliplr(patch_block), block_size,
-                                                       overlap)
-        mask_right = np.fliplr(mask_right)
-        mask_right = blur_patch_mask(mask_right, block_size, overlap, has_left, has_right, has_top, has_bottom)
-        masks_list.append(mask_right)
-
-    if has_top:
-        # V , >  counterclockwise rotation
-        mask_top = get_min_cut_patch_mask_horizontal(np.rot90(ref_block_top), np.rot90(patch_block), block_size, overlap)
-        mask_top = np.rot90(mask_top, 3)
-        mask_top = blur_patch_mask(mask_top, block_size, overlap, has_left, has_right, has_top, has_bottom)
-        masks_list.append(mask_top)
-
-    if has_bottom:
-        mask_bottom = get_min_cut_patch_mask_horizontal(np.fliplr(np.rot90(ref_block_bottom)),
-                                                        np.fliplr(np.rot90(patch_block)), block_size, overlap)
-        mask_bottom = np.rot90(np.fliplr(mask_bottom), 3)
-        mask_bottom = blur_patch_mask(mask_bottom, block_size, overlap, has_left, has_right, has_top, has_bottom)
-        masks_list.append(mask_bottom)
-
-    # --- apply masks and return block ---
-    # compute auxiliary data
-    mask_s = sum(masks_list)
-    masks_max = np.maximum.reduce(masks_list)
-    mask_mos = np.divide(masks_max, mask_s, out=np.zeros_like(mask_s), where=mask_s != 0)
-    # note -> if blurred, masks weights are scaled with respect the max mask value.
-    # example: mask1: 0.2 mask2:0.5 -> max = .5 ; sum = .7 -> (.2*lb + .5*tb) * (max/sum) + patch * (1 - max)
-
-    # place adjacent block sections
-    res_block = np.zeros_like(patch_block)
-    if has_left:
-        res_block += mask_left * np.roll(ref_block_left, overlap, 1)
-    if has_right:
-        res_block += mask_right * np.roll(ref_block_right, -overlap, 1)
-    if has_top:
-        res_block += mask_top * np.roll(ref_block_top, overlap, 0)
-    if has_bottom:
-        res_block += mask_bottom * np.roll(ref_block_bottom, -overlap, 0)
-    res_block *= mask_mos
-
-    # place patch section
-    patch_weight = 1 - masks_max
-    res_block = res_block + patch_weight * patch_block
-
-    return res_block
 
 
 def make_seamless_vertically(image, block_size, overlap, tolerance, rng: np.random.Generator,
